requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9

